cros: Avoid double-loading the EC image
At present the EC image is in the same area as the U-Boot binary and is
therefore loaded as part of RW firmware. We can simple copy it from its
resulting memory location instead of loading it again from SPI flash.

This option is controlled by the device tree.

BUG=chrome-os-partner:13961
BRANCH=link
TEST=manual
Boot through RW to kernel on link
Update the EC and see that a software sync happens (only once).

Change-Id: I12efa56a620a6656c5c54bfae145c0bfc659d2a0
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/40128
Reviewed-by: Randall Spangler <rspangler@chromium.org>