cros: Permit compression of blobs
On some platforms SPI flash is so excruciatingly slow that we want to
compress our firmware. Add support for this, controlled by a 'compress'
property in the fdt node.

The only compression currently supported is lzo.

BUG=chrome-os-partner:13961
BRANCH=link
TEST=manual
With compression enabled for link on RW A and B, boot through RW to
kernel
Perform test again with compression disabled on both RW A and B.

Change-Id: Ieb60d8454bafad1cf840232b9f0f84ce0ac020bd
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/40129
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>