x86: Enable bootstage for coreboot
This is a convenient way of finding out where boottime is going. Enable
it for coreboot.

BUG=chrome-os-partner:13961
BRANCH=link
TEST=manual
Boot through firmware and see a bootstage report on the console just before
the kernel is run. This appears both when booting RO and RO+RW firmware,
but only once in each case. For the RO+RW case, the timings for both RO and
RW are sorted and shown together. Coreboot timings are also show and mesh
properly with the time base used by U-Boot.

Change-Id: I24e5345a5a6023a1a2f5fe6bb253c9d16c89835a
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/40126