cros: For x86, use the top of the FDT as the end of U-Boot's space
The old code adds a memory-clearing hole for U-Boot's code and data.
We place the FDT at the top of RAM, so need to avoid clearing it. Since
the FDT is above U-Boot's code and data (as per coreboot's sdram.c), it is
enough for the memory-clearing hole to extend from the start of U-Boot's
code to the end of the FDT.

BUG=chrome-os-partner:13961
BRANCH=link
TEST=manual
build and boot to kernel on link
build to recovery and see that we no longer get FDT errors while waiting for
media. The error was:

vboot_flag_sysinfo.c: vboot_flag_fetch_sysinfo: failed to read name of vboot-flag-lid-open

Change-Id: I5d47ec870195b5fcc6b901e4ede8ea74e782db96
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/40111
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Brian Harring <ferringb@chromium.org>