exynos: video: use the lane count and link rate from the DPCD
Use the lane count and the link extracted from the DPCD instead of
forcing the hardcoded value in the source file.
Now, the DP initialization code works on bridges with either 1 or 2 data
lanes with modifying the source or the configuration.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=none
BUG=chrome-os-partner:14389
TEST=on both Snow and Spring, boot the platform with a non verified bootloader
and see the firmware screen coming up.

Change-Id: Ie00be183c90391118c5e5474a6b6e38e10d05a6f
Reviewed-on: https://gerrit.chromium.org/gerrit/41456
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>