exynos: re-enumerate properly on HSIC hub
When doing several "usb start" in a row, the subsequent ones were no
longer enumerating a USB key connected to a USB2.0 port of the HSIC hub.

We need to keep the HSIC hub and controller under reset until we
initialize the interface to ensure we are bringing up both sides in the
right order and within the right timings.
To achieve this, just shutdown properly the on-board PHY and the external
chip when stopping the USB stack.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=none
BUG=chrome-os-partner:17535
TEST=On Spring with a USB key plugged, at U-Boot prompt, type "usb
start" several times in a row and see everytime:
"scanning bus for storage devices... 1 Storage Device(s) found"

Change-Id: I9c969347119372e4abd67dac24427a5f2428b054
Reviewed-on: https://gerrit.chromium.org/gerrit/42469
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Jeremy Thorpe <jeremyt@chromium.org>