gen: usb: Add new command to set USB 2.0 port test modes
This patch adds a new 'usb test' command, that will set a port to a USB
2.0 test mode (see USB 2.0 spec 7.1.20). It supports all five test modes
on both downstream hub ports and ordinary device's upstream ports. In
addition, it supports EHCI root hub ports.

BUG=chrome-os-partner:16153
TEST=1.Connect a USB analyzer to a Snow's 2.0 port and run 'usb test 1 1
Packet'. Ensure that the analyzer shows a constant stream of test
packets.
2. Connect a USB analyzer to a Spring's left port and run 'usb test 2 1
Packet'. Ensure that the analyzer shows a constant stream of test
packets.

Change-Id: I5f59548e4f0bcd35992680d2a86bef51b3eacfb4
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/43420
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>