DO NOT SUBMIT: environment updates for booting regular distro
Take out a bunch of the very verbose and elaborate boot options, and go
for something much much simpler for regular distro booting.

The default args try to boot uImage from partition 12 on the eMMC, which
is the legacy syslinux partition.

DO NOT SUBMIT. This is just for out-of-tree usage. Uploading only to
share with others.

Change-Id: I1d05baf70c83afab35e9ca3cf8a2798ae2776608
Signed-off-by: Olof Johansson <olofj@chromium.org>
Signed-off-by: Nikolay Nikolaev <n.nikolaev@virtualopensystems.com>