The exynos5250 chip is affected by the errata
  773022 (disable loop buffer)
  774769 (refrain streaming-write to ever allocate into the L2 cache)

and needs to get the respective fixes as mentioned above applied.

This is done in the Samsung downstream kernel within Linux. However, that
approach fails to work when we run the guest kernel in HYP mode, as in
now Linux's errata fixup code runs without privileges to access the
aux control register.

So instead, let's patch up u-boot (where CPU errata fixes belong to
anyways), so that we can happily run a kernel.

This patch originates from Alexander Graf <agraf at suse.de>

Tested-by: Nikolay Nikolaev <n.nikolaev@virtualopensystems.com>